refactor: separated qr function in a different file called QR following SOLID principle
