+++ src/telegram_py_bot.py
@@ -1,20 +1,5 @@
 from telegram.ext import ApplicationBuilder, CommandHandler, MessageHandler, filters
-import qrcode
-
-# qr code image generation function
-def QR(text):
-    qr = qrcode.QRCode(
-        version=1,
-        error_correction=qrcode.constants.ERROR_CORRECT_L,
-        box_size=15,
-        border=4,
-    )
-
-    qr.add_data(text)
-    qr.make(fit=True)
-
-    img = qr.make_image(fill_color="black", back_color="white")
-    img.save("qrcode.png")
+import QR
 
 # telegram bot start command handler
 async def start(update, context):
